+++ hystrix/src/main/java/feign/hystrix/HystrixFeign.java
@@ -46,8 +46,16 @@ public final class HystrixFeign {
     /**
      * @see #target(Class, String, Object)
      */
-    public <T> T target(Target<T> target, final T fallback) {
-      return buildWithFallback(fallback).newInstance(target);
+    public <T> T target(Target<T> target, T fallback) {
+      return build(fallback != null ? new FallbackFactory.Default<T>(fallback) : null)
+          .newInstance(target);
+    }
+
+    /**
+     * @see #target(Class, String, FallbackFactory)
+     */
+    public <T> T target(Target<T> target, FallbackFactory<? extends T> fallbackFactory) {
+      return build(fallbackFactory).newInstance(target);
     }
 
     /**
@@ -89,6 +97,14 @@ public final class HystrixFeign {
       return target(new Target.HardCodedTarget<T>(apiType, url), fallback);
     }
 
+    /**
+     * Same as {@link #target(Class, String, T)}, except you can inspect a source exception before
+     * creating a fallback object.
+     */
+    public <T> T target(Class<T> apiType, String url, FallbackFactory<? extends T> fallbackFactory) {
+      return target(new Target.HardCodedTarget<T>(apiType, url), fallbackFactory);
+    }
+
     @Override
     public Feign.Builder invocationHandlerFactory(InvocationHandlerFactory invocationHandlerFactory) {
       throw new UnsupportedOperationException();
@@ -102,15 +118,15 @@ public final class HystrixFeign {
 
     @Override
     public Feign build() {
-      return buildWithFallback(null);
+      return build(null);
     }
 
     /** Configures components needed for hystrix integration. */
-    Feign buildWithFallback(final Object nullableFallback) {
+    Feign build(final FallbackFactory<?> nullableFallbackFactory) {
       super.invocationHandlerFactory(new InvocationHandlerFactory() {
         @Override public InvocationHandler create(Target target,
             Map<Method, MethodHandler> dispatch) {
-          return new HystrixInvocationHandler(target, dispatch, setterFactory, nullableFallback);
+          return new HystrixInvocationHandler(target, dispatch, setterFactory, nullableFallbackFactory);
         }
       });
       super.contract(new HystrixDelegatingContract(contract));